General refactors for terraform
Some general refactors which didn't fit into a ticket.

- Set ConsoleJSON to false from code also.
- Moved the runCommand method to a separate file as everything was being
shoved into create.go.
- Changed the updateConfig method to return an error and perform
logging only from Create.
- Removed some TODOs
- Added some logging which gives an indication as to what is happening
after a terraform deployment finishes because there's still some things
that are done from inside the code.
- Increased the command timeout to 10 minutes because RDS deployment
sometimes takes close to 4.40 minutes and when we move to cluster
deployment, it's sure going to take a lot more.
- Replaced errors.Wrap with fmt.Errorf which is used more consistently.
